Do not use C-style casts

+++ src/ui/SourceViewerWidget.cpp
@@ -214,8 +214,8 @@ void MarginWidget::paintEvent(QPaintEvent *event)
 	painter.fillRect(event->rect(), Qt::transparent);
 
 	QTextBlock block = m_sourceViewer->firstVisibleBlock();
-	int top = (int) m_sourceViewer->blockBoundingGeometry(block).translated(m_sourceViewer->contentOffset()).top();
-	int bottom = (top + (int) m_sourceViewer->blockBoundingRect(block).height());
+	int top = m_sourceViewer->blockBoundingGeometry(block).translated(m_sourceViewer->contentOffset()).top();
+	int bottom = (top + m_sourceViewer->blockBoundingRect(block).height());
 	const int right = (width() - 5);
 	const int selectionStart = m_sourceViewer->document()->findBlock(m_sourceViewer->textCursor().selectionStart()).blockNumber();
 	const int selectionEnd = m_sourceViewer->document()->findBlock(m_sourceViewer->textCursor().selectionEnd()).blockNumber();
@@ -233,7 +233,7 @@ void MarginWidget::paintEvent(QPaintEvent *event)
 
 		block = block.next();
 		top = bottom;
-		bottom = (top + (int) m_sourceViewer->blockBoundingRect(block).height());
+		bottom = (top + m_sourceViewer->blockBoundingRect(block).height());
 	}
 }
